# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RatelPlus cross-system contracts
  version: 0.1.0-draft
  description: |
    The two seams where RatelBSS touches the network (Build Plan, "The two contracts to freeze on
    day one"). This file is the source of truth. If code and this file disagree, the code is wrong.

    - Contract 1: RatelLink (tag `RatelLink`), runs on core-cp.
    - Contract 2: RatelMeter (tag `RatelMeter`), API runs on bss-app.

    Conventions from the Build Plan:
    - Plain REST, standard HTTP status codes, JSON bodies.
    - Errors: `{"error": {"code": "...", "message": "..."}}`.
    - Every call carries an API key in the `Authorization` header, one key per calling system.
    - Every call that changes something accepts an `Idempotency-Key` header. Repeating a call with
      the same key returns the first result.
    - Times are Unix epoch seconds in UTC, sent as integers. Usage is in bytes.
    - Until RatelLink's live changes land, every change takes effect at the line's next attach.
    - Ki and OPc are write-only. No response in this file may contain them.

    Anything the Build Plan does not specify is marked `TODO(contract)` in a description or with
    the `x-todo` extension. Do not fill those in by guessing. Agree them with the owners, then
    update this file in the same commit as the code.

    Not covered here on purpose: RatelBSS's own API for RatelDesk/RatelPay, and the internal
...
  x-todo:
...
servers:
  - url: http://127.0.0.1:4010
...
security:
  - ApiKeyAuth: []
...
paths:
  /v1/sims:
    post:
      tags: [RatelLink]
      operationId: link_create_sim
      summary: Store a SIM's keys in RatelLink. No service yet.
      description: |
        Stores the SIM's keys in RatelLink. Does not create service. RatelBSS never sends this:
        keys go to RatelLink and RatelBSS receives only IMSI and ICCID.
        Idempotent: sending the same keys again for an IMSI succeeds and changes nothing. An IMSI
        that already has DIFFERENT keys is refused with 409 and the stored keys are never replaced.
        TODO(contract): the Build Plan's sim_key store also lists `amf`; the request body here
        lists only imsi, ki, opc as in Contract 1.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimImport"
      responses:
        "200":
          $ref: "#/components/responses/OkEmpty"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/UnexpectedError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        One API key per calling system, rotated every 90 days. Keys never live in the repository.
        Sent as `Authorization: Bearer <key>`. A missing or invalid key gets 401.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Repeating a call with the same key returns the first result.
        TODO(contract): key format, length and retention window.
      schema:
        type: string
        minLength: 1
...
  responses:
    OkEmpty:
      description: |
        Success. TODO(contract): response body is not specified in the Build Plan.
      content:
        application/json:
          schema:
            type: object
    Unauthorized:
      description: Missing or invalid Bearer API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The IMSI already has different keys stored. Nothing was changed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: |
        Invalid request. TODO(contract): confirm 422 versus 400 for validation failures.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnexpectedError:
      description: Any other failure, in the standard error shape.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
    Imsi:
      type: string
      pattern: "^[0-9]{15}$"
      description: The SIM's 15 digit identity on the network. Ours start with 62100.
      examples: ["621000000000001"]
    SimImport:
      type: object
      required: [imsi, ki, opc]
      additionalProperties: false
      properties:
        imsi:
          $ref: "#/components/schemas/Imsi"
        ki:
          type: string
          pattern: "^[0-9a-fA-F]{32}$"
          writeOnly: true
          description: Secret. Write-only. Never returned, logged or sent to RatelBSS.
          x-todo: Not in the Build Plan. 128 bits as 32 hex characters is an assumption to confirm with the SIM supplier.
        opc:
          type: string
          pattern: "^[0-9a-fA-F]{32}$"
          writeOnly: true
          description: Secret. Write-only. Never returned, logged or sent to RatelBSS.
          x-todo: Not in the Build Plan. 128 bits as 32 hex characters is an assumption to confirm with the SIM supplier.
...